requirements.txt
scipy == 1.1.0
matplotlib >= 2.2.2
numpy >= 1.15.0
progressbar2 == 3.38.0
